+++ dcmimgle/include/diovlay.h
@@ -22,9 +22,9 @@
  *  Purpose: DicomOverlay (Header)
  *
  *  Last Update:      $Author: joergr $
- *  Update Date:      $Date: 1998-12-23 11:36:28 $
+ *  Update Date:      $Date: 1998-12-23 13:21:29 $
  *  Source File:      $Source: /export/gitmirror/dcmtk-git/../dcmtk-cvs/dcmtk/dcmimgle/include/Attic/diovlay.h,v $
- *  CVS/RCS Revision: $Revision: 1.6 $
+ *  CVS/RCS Revision: $Revision: 1.7 $
  *  Status:           $State: Exp $
  *
  *  CVS/RCS Log at end of file
@@ -158,8 +158,8 @@ class DiOverlay
     int addPlane(const unsigned int group,
                  const signed int left,
                  const signed int top,
-                 const unsigned long columns,
-                 const unsigned long rows,
+                 const unsigned int columns,
+                 const unsigned int rows,
                  const DcmOverlayData &data,
                  const DcmLongString &label,
                  const DcmLongString &description,
@@ -241,7 +241,10 @@ class DiOverlay
  *
  * CVS/RCS Log:
  * $Log: diovlay.h,v $
- * Revision 1.6  1998-12-23 11:36:28  joergr
+ * Revision 1.7  1998-12-23 13:21:29  joergr
+ * Changed parameter type (long to int) to avoid warning reported by MSVC5.
+ *
+ * Revision 1.6  1998/12/23 11:36:28  joergr
  * Changed order of parameters for addOverlay() and getOverlayData().
  * Corrected bug concerning additional overlay planes.
  *
